Revert "Extrapolate NotificationDot to components"
This reverts commit 5c8c8dca410a434af9942fec7f5438ce4a8dca8c.
