--- src/components/NotificationDot.tsx
@@ -1,28 +1,0 @@
-/*
- * Vencord, a Discord client mod
- * Copyright (c) 2023 Vendicated and contributors
- * SPDX-License-Identifier: GPL-3.0-or-later
- */
-
-import { classes } from "@utils/misc";
-import { findByCodeLazy,findByPropsLazy } from "@webpack";
-
-const getDotWidth = findByCodeLazy("<10?16:");
-const dotStyles = findByPropsLazy("numberBadge");
-
-
-const NotificationDot = ({ unreadCount, mentionCount }: { unreadCount: number; mentionCount: number; }) => {
-    return unreadCount > 0 ?
-        <div
-            data-has-mention={!!mentionCount}
-            className={classes(dotStyles.numberBadge, dotStyles.baseShapeRound)}
-            style={{
-                backgroundColor: mentionCount ? "var(--status-danger)" : "var(--brand-experiment)",
-                width: getDotWidth(mentionCount || unreadCount)
-            }}
-        >
-            {mentionCount || unreadCount}
-        </div > : null;
-};
-
-export default NotificationDot;
